Conflicts:
	grails-datastore-gorm-cassandra/src/main/groovy/org/grails/datastore/mapping/cassandra/CassandraDatastore.java

+++ grails-datastore-gorm-cassandra/src/test/groovy/org/grails/datastore/mapping/cassandra/BasicPersistenceSpec.groovy
@@ -3,7 +3,7 @@ package org.grails.datastore.mapping.cassandra
 import grails.gorm.tests.GormDatastoreSpec
 import grails.gorm.tests.TestEntity
 
-import org.grails.datastore.mapping.cassandra.utils.UUIDUtil;
+import org.grails.datastore.mapping.cassandra.utils.UUIDUtil
 
 /**
  * @author Graeme Rocher
@@ -12,42 +12,41 @@ import org.grails.datastore.mapping.cassandra.utils.UUIDUtil;
 class BasicPersistenceSpec extends GormDatastoreSpec {
 
     void testBasicPersistenceOperations() {
-        when:
-        def te = session.retrieve(TestEntity, UUIDUtil.getRandomTimeUUID())
+        when: "read non existent"
+        	def te = session.retrieve(TestEntity, UUIDUtil.getRandomTimeUUID())
 
         then:
             te == null
         
-        when:
-            te = new TestEntity(name: "Bob", age: 45)
-    
+        when: "save and flush"
+            te = new TestEntity(name: "Bob", age: 45)    
             session.persist(te)
             session.flush()
+			
         then:
             te != null
             te.id != null
             te.id instanceof UUID
     
-        when:
+        when: "read"
             session.clear()
             def t2 = session.retrieve(TestEntity, te.id)
     
-        then:
-    
+        then:    
             t2 != null
             t2.name == "Bob"
             t2.age == 45
             t2.id != null
             t2.id instanceof UUID
         
-        when:
+        when: "update no flush"
             te.age = 55
             session.persist(te)
             def tcached = session.retrieve(TestEntity, te.id)
         then:
             tcached == te
         
-        when:    
+        when: "flush and persist to db and read again"    
             session.flush()
             session.clear()             
             te = session.retrieve(TestEntity, te.id)
@@ -58,7 +57,7 @@ class BasicPersistenceSpec extends GormDatastoreSpec {
             te.name == "Bob"
             te.age == 55
         
-        when:
+        when: "delete"
             session.delete(te)
             session.flush()            
             def deletedEntity = session.retrieve(TestEntity, te.id)
